src/requirements.txt
cloudshell-core>=2.2,<2.3
cloudshell-shell-core>=3.1,<3.2
cloudshell-networking-devices>=2.1,<2.2
cloudshell-sdn-odl>=1.0,<1.1